Refactored room edges
